Abstract PokemonID operations
